@@ -21,7 +21,7 @@
         }
     ],
     "require": {
-        "php": ">=5.4",
+        "php": ">=5.6",
         "lstrojny/functional-php": "dev-master",
         "internations/solr-utils": "~0.4"
     },
