@@ -6,7 +6,7 @@ class ParameterExpression extends Expression
     /**
      * @var array
      */
-    private $parameters = array();
+    private $parameters = [];
 
     /**
      * @param array $parameters
@@ -21,9 +21,9 @@ class ParameterExpression extends Expression
      */
     public function __toString()
     {
-        $parameters = array_map(array($this, 'replaceNull'), $this->parameters);
+        $parameters = array_map([$this, 'replaceNull'], $this->parameters);
 
-        return implode(', ', array_map(array('InterNations\Component\Solr\Util', 'sanitize'), $parameters));
+        return implode(', ', array_map(['InterNations\Component\Solr\Util', 'sanitize'], $parameters));
     }
 
     private function replaceNull($value)
